+++ project/Starfighter/include/game/Supernova.h
@@ -10,7 +10,7 @@ class Supernova : public Obstacle
 {
     //Q_OBJECT
 public:
-    Supernova(int x, int y,GameEngine* ge);
+    Supernova(qreal x, qreal y,GameEngine* ge);
     ~Supernova();
 
     QRectF boundingRect() const;
