+++ project/Starfighter/src/game/ProjectileV.cpp
@@ -16,7 +16,7 @@
 
 #define INTERVAL_X 3
 
-ProjectileV::ProjectileV(int _originX, int _originY,Shooter _from, qreal _ampli, qreal _omega)
+ProjectileV::ProjectileV(qreal _originX, qreal _originY,Shooter _from, qreal _ampli, qreal _omega)
     :Displayable(_originX,_originY),Projectile(_originX,_originY,_from),dAmplitude(_ampli),dOmega(_omega)
 {
     power = POWER_V;
