+++ project/Starfighter/src/game/Spaceship.cpp
@@ -14,7 +14,7 @@
 #define AMPLI_PV 60.0
 #define OMEGA_PV 1.0
 
-Spaceship::Spaceship(int x,int y,Shooter _player,GameEngine *ge):Displayable(x,y),player(_player),gameEngine(ge)
+Spaceship::Spaceship(qreal x,qreal y,Shooter _player,GameEngine *ge):Displayable(x,y),player(_player),gameEngine(ge)
 {
     speed = SPEED_DEFAULT;
     type = ProjCross;
